requirements.txt
streamlit
snowflake-snowpark-python
fpdf
pandas
markdown2
bs4